improve responses - phase2

+++ api/src/main/java/com/api/common/dto/EstimationStats.java
@@ -3,5 +3,5 @@ package com.api.common.dto;
 public record EstimationStats(
   double total,
   double average,
-  long count
+  int count
 ) {}
